+++ CHANGELOG.md
@@ -5,12 +5,14 @@ All notable changes to this project will be documented in this file. The format 
 
 ### Added
 - Add plot- and screenshot-methods to `Region` and `Scheme` (base class for quadratures).
+- Add `item = FormItem(bilinearform, linearform=None)` to be used as an item in a `Step(items=[item])`.
 
 ### Changed
 - Refactor the assembly-submodule. Move the weak-form expression-related classes to the `assembly.expression` submodule.
 - Move `Basis` to the new `assembly.expression` submodule.
 - Make the `field`-submodule public.
 - Always `import felupe as fem` in docs and tests.
+- Change default optional (keyword) arguments of a weak-form expression decorator from `Form(args=(), kwargs={})` to `Form(args=None, kwargs=None)`.
 
 ### Removed
 - Don't import `Basis` to the global namespace (not necessary as it is used only internally by the weak-`Form` expression decorator).